+++ document/cloud-client/parse_form_beta_test.py
@@ -13,14 +13,14 @@
 # limitations under the License.
 
 import os
-import parse_form
+import parse_form_beta
 
 PROJECT_ID = os.environ['GCLOUD_PROJECT']
 INPUT_URI = 'gs://cloud-samples-data/documentai/form.pdf'
 
 
 def test_parse_form(capsys):
-    parse_form.parse_form(PROJECT_ID, INPUT_URI)
+    parse_form_beta.parse_form(PROJECT_ID, INPUT_URI)
     out, _ = capsys.readouterr()
     assert 'Field Name' in out
     assert 'Field Value' in out
