+++ document/cloud-client/quickstart_beta.py
